Updated b-tree.c
Removed the memory allocation for the B-tree root.

Removed unnecessary if condition in free_BT().

Added root memory allocation logic to insert_BT(), including simple insertion due to its property as a root.

In search_BT(), added a variable key_count, along with a comment that explains why. Also added an if statement to the value assignment to variable page after the for loop.